pyproject.toml
# see documentation, e.g.
# - https://packaging.python.org/en/latest/specifications/declaring-project-metadata/#declaring-project-metadata
# - https://setuptools.pypa.io/en/latest/userguide/pyproject_config.html
# - https://www.python.org/dev/peps/pep-0621/

[build-system]
requires = ["setuptools>=64.0.0", "setuptools-scm", "wheel"]
build-backend = "setuptools.build_meta"

[project]
authors = [
    { name = "Nicolas Renaud", email = "n.renaud@esciencecenter.nl" },
    { name = "TNO Quantum", email = "tnoquantum@tno.nl" },
]
maintainers = [
    { name = "Nicolas Renaud", email = "n.renaud@esciencecenter.nl" },
    { name = "TNO Quantum", email = "tnoquantum@tno.nl" },
]
classifiers = [
    "Development Status :: 2 - Pre-Alpha",
    "Intended Audience :: Developers",
    "Intended Audience :: Science/Research",
    "Intended Audience :: Information Technology",
    "License :: OSI Approved :: Apache Software License",
    "Natural Language :: English",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Private :: Do Not Upload to pypi server",
    "Operating System :: OS Independent",
    "Typing :: Typed",
]
dependencies = [
    "numpy>=1.21.6",
    "networkx>=3.2.1",
    "matplotlib",
    "dimod",
    "dwave-samplers",
    "dwave-system",
    "scipy==1.11.4",
    "pads @ git+https://github.com/QuantumApplicationLab/PADS",
]
description = "Reoder matrix elements to minimize LU fill-ins"
keywords = ["matrix", "graph", "quantum", "QUBO", "benders cuts"]
license = { file = "LICENSE" }
name = "qreorder"
readme = { file = "README.md", content-type = "text/markdown" }
requires-python = ">=3.8"
dynamic = ["version"]

[project.optional-dependencies]
dev = [
    "build",            # build is not only used in publishing (below), but also in the template's test suite
    "bump-my-version",
    "coverage [toml]",
    "pytest",
    "pytest-cov",
    "ruff",
    "sphinx",
    "sphinx_rtd_theme",
    "sphinx-autoapi",
    "tox",
    "myst_parser",
    "black",
    "isort",
    "pylint",
    "mypy",
]
publishing = ["build", "twine", "wheel"]

[project.urls]
Repository = "https://github.com/QuantumApplicationLab/qreorder"
Issues = "https://github.com/QuantumApplicationLab/qreorder/issues"
Changelog = "https://github.com/QuantumApplicationLab/qreorder/CHANGELOG.md"


[tool.pytest.ini_options]
testpaths = ["tests"]

[tool.coverage.run]
branch = true
source = ["qreorder"]
command_line = "-m pytest"

[tool.isort]
lines_after_imports = 2
force_single_line = 1
no_lines_before = [
    "FUTURE",
    "STDLIB",
    "THIRDPARTY",
    "FIRSTPARTY",
    "LOCALFOLDER",
]
known_first_party = "qreorder"
src_paths = ["qreorder", "tests"]
line_length = 88

# For completeness, until we move to an src-based layout
[tool.setuptools.packages.find]
include = ["qreorder*"]
exclude = ["tests*"]

[tool.tox]
legacy_tox_ini = """
[tox]
envlist = py38,py39,py310,py311,py312
skip_missing_interpreters = true
[testenv]
commands = pytest
extras = dev
"""

[tool.ruff]
# Enable Pyflakes `E` and `F` codes by default.
select = [
    "F", # Pyflakes
    "E", # pycodestyle (error)
    "W", # pycodestyle (warning)
    # "C90",  # mccabe
    "I", # isort
    "D", # pydocstyle
    # "PL",   # Pylint
    # "PLC",  # Convention
    # "PLE",  # Error
    # "PLR",  # Refactor
    # "PLW",  # Warning

]
ignore = [
    'D100', # Missing module docstring
    'D104', # Missing public package docstring
    # The following list excludes rules irrelevant to the Google style
    'D203',
    'D204',
    'D213',
    'D215',
    'D400',
    'D401',
    'D404',
    'D406',
    'D407',
    'D408',
    'D409',
    'D413',
]

# Allow autofix for all enabled rules (when `--fix`) is provided.
fixable = ["A", "B", "C", "D", "E", "F", "I"]
unfixable = []

exclude = [
    ".bzr",
    ".direnv",
    ".eggs",
    ".git",
    ".hg",
    ".mypy_cache",
    ".nox",
    ".pants.d",
    ".ruff_cache",
    ".svn",
    ".tox",
    ".venv",
    "__pypackages__",
    "_build",
    "buck-out",
    "build",
    "dist",
    "node_modules",
    "venv",
    ".venv",
    "scripts",
]
per-file-ignores = {}


# Allow unused variables when underscore-prefixed.
dummy-variable-rgx = "^(_+|(_+[a-zA-Z0-9_]*[a-zA-Z0-9]+?))$"

target-version = "py39"
line-length = 120

[tool.ruff.isort]
known-first-party = ["qreorder"]
force-single-line = true
no-lines-before = [
    "future",
    "standard-library",
    "third-party",
    "first-party",
    "local-folder",
]

[tool.bumpversion]
current_version = "0.1.0"

[[tool.bumpversion.files]]
filename = "qreorder/__init__.py"

[[tool.bumpversion.files]]
filename = "pyproject.toml"

[[tool.bumpversion.files]]
filename = "CITATION.cff"

[[tool.bumpversion.files]]
filename = "docs/conf.py"

[tool.mypy]
plugins = ["numpy.typing.mypy_plugin"]
strict = true
show_error_context = true
install_types = true
explicit_package_bases = true
namespace_packages = true
show_error_codes = true
ignore_missing_imports = true

[tool.pylint."TYPECHECK"]
generated-members = ["torch.*", "numpy.*"]

[tool.pylint.similarities]
ignore-imports = "yes"
ignore-signatures = "yes"
